refactor(website): render the homepage shell in Astro
